requirements.txt
edge-tts>=6.1.0
moviepy>=2.0.0
Pillow>=10.0.0
requests>=2.28.0
numpy>=1.24.0
